+++ pkg/parsergen/backend/golangtable/doc.go
@@ -1,0 +1,3 @@
+// Package golangtable provides a parser generator backend which takes a parser as input and outputs Go source code
+// which keeps the parse table in lookup tables.
+package golangtable